Fleshed out local storage and reorg of project
Needs to mock services for unit testing, and finish load of initial state from local storage
