--- BuffTracker/BuffTracker.App/Models/NotifyPropertyChangedBase.cs
@@ -1,24 +1,0 @@
-﻿using System.ComponentModel;
-using System.Runtime.CompilerServices;
-
-namespace BuffTracker.App.Models;
-
-public abstract class NotifyPropertyChangedBase : INotifyPropertyChanged
-{
-    public event PropertyChangedEventHandler? PropertyChanged;
-    protected void NotifyPropertyChanged([CallerMemberName] string propertyName = null)
-    {
-        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-    }
-
-    protected bool SetProperty<T>(ref T storage, T value, [CallerMemberName] string propertyName = null)
-    {
-        if (Equals(storage, value))
-        {
-            return false;
-        }
-        storage = value;
-        NotifyPropertyChanged(propertyName);
-        return true;
-    }
-}
