+++ BuffTracker/BuffTracker.App/Program.cs
@@ -1,7 +1,7 @@
 using Blazored.LocalStorage;
 using BuffTracker.App;
-using BuffTracker.App.Services;
-using BuffTracker.App.ViewModels;
+using BuffTracker.Shared.Services;
+using BuffTracker.ViewModels;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 
@@ -10,8 +10,8 @@ builder.RootComponents.Add<App>("#app");
 builder.RootComponents.Add<HeadOutlet>("head::after");
 
 builder.Services.AddSingleton<BuffTrackerViewModel>();
-builder.Services.AddSingleton<BuffTrackerState>();
-builder.Services.AddSingleton<TimelineEngine>();
+builder.Services.AddSingleton<AppState>();
+builder.Services.AddSingleton<TimelineViewModel>();
 builder.Services.AddSingleton<StatusRulesEngine>();
 
 builder.Services.AddBlazoredLocalStorageAsSingleton();
