+++ BuffTracker/BuffTracker.App/_Imports.razor
@@ -9,6 +9,3 @@
 @using BuffTracker.App
 @using BuffTracker.App.Shared
 @using BuffTracker.App.Components
-@using BuffTracker.App.Models
-@using BuffTracker.App.Services
-@using BuffTracker.App.ViewModels
