+++ BuffTracker/BuffTracker.ViewModels/BuffTracker.ViewModels.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BuffTracker.Shared\BuffTracker.Shared.csproj" />
+  </ItemGroup>
+
+</Project>
